move dependencies api to deps
